add meanvalue

+++ src/main/scala/org/grakovne/organic/client/messages/Measurement.scala
@@ -1,19 +1,19 @@
 package org.grakovne.organic.client.messages
 
-final case class Measurement(tvoc: Int, carbonDioxide: Int, temperature: Double, humidity: Double)
+final case class Measurement(tvoc: Double, carbonDioxide: Double, temperature: Double, humidity: Double)
 
 class MeasurementBuilder {
-  private var tvoc: Option[Int] = Option.empty[Int]
-  private var carbonDioxide: Option[Int] = Option.empty[Int]
+  private var tvoc: Option[Double] = Option.empty[Double]
+  private var carbonDioxide: Option[Double] = Option.empty[Double]
   private var temperature: Option[Double] = Option.empty[Double]
   private var humidity: Option[Double] = Option.empty[Double]
 
-  def insertTvoc(tvoc: Option[Int]): MeasurementBuilder = {
+  def insertTvoc(tvoc: Option[Double]): MeasurementBuilder = {
     this.tvoc = tvoc
     this
   }
 
-  def insertCarbonDioxide(carbonDioxide: Option[Int]): MeasurementBuilder = {
+  def insertCarbonDioxide(carbonDioxide: Option[Double]): MeasurementBuilder = {
     this.carbonDioxide = carbonDioxide
     this
   }
